Merge 0.9-clusterdb branch back into the trunk, adding things like (thanks Bull ;-): * remove command * tune command * quota support * ClusterDB backend * improved Backend API
Also the following points have been modified during merging:
* added compatibility for ClusterShell 1.0 RC3
* available modes are now : File, ClusterDB (case sensitive)
  and dependency to ClusterDB or File mode has been removed
* as the tag is not always defined in File mode, it is no
  more used directly in format/status
  (replaced by Target.get_id())
* the XMF file is removed in Configuration.FileSystem
* changed exceptions in TuningModel to inherits ConfigException
* modified node_list by ClusterShell.NodeSet
* changed indentation to fit in a 80-columns terminal
  (identation of Backend.ClusterDB NOT modified)
* added BACKEND_MODNAME="ClusterDB" to ClusterDB.py

Work well in File mode, to be tested in ClusterDB mode.
Be careful to change backend=clusterdb to backend=ClusterDB,
or backend=file to backend=File. This should be fixed in a
elegant way in a near future.

The 0.9-clusterdb branch should be closed and a new
one can be created from the trunk.

SVN commit: r62