Ticket #113: Reduce entropy with target in backend.
Target list returned by backends could be in any order. If you
remove/install the same filesystem in the same conditions, you would
like to generate the same filesystem. To be sure order will be the same,
we sort them. This also helps for filesystem update/comparison.

SVN commit: r378