Ticket #25: Clean Proxy with output/err management.
Incoming messages are either Shine message directly handled or something
else which is keep in a dedicated MsgTree, which is used at proxy
command end to trigger error messages if needed. This avoid filtering the
task MsgTree and the pickle message in it.

The nodes which failed without returning any message has a better
handling too.