Ticket #103: Identical error messages are grouped.
Proxy error messages are managed by a MsgTree which provides
features to easily group output messages.
Hostnames in message output are modified. THIS_SHINE_HOST pattern
is used for that. Hopefully it will not be used in regular output.
When displayed, THIS_SHINE_HOST pattern is replaced by the
corresponding hostnames.