Ticket #174: Better error handling when reading remote messages
It could be sometimes impossible to read messages sent by remote shine
instances. Most of the time, this is due to unpickling errors because
Shine or ClusterShell version mismatch between the admin node and the
remote nodes.

There is better detection of theses cases and a better reporting of this
problem.