Ticket #186: Fix return code for stop/umount
If action failed when stopping or unmounting the component stayed
online and the return code was always '0'. This is now fixed.

Ticket #186 could be improved later, unmounting using the mount
path and/or checking expected state again at ev_close in case
of success.

Also fix or add missing lines.

Change-Id: Ifda5b2422e8f48e1561a182af5b33f8d5fc606cf