Ticket #46: Call Install() in _prepare for start,mount,tune
Install, responsible for copying tuning.conf before running commands,
is now run inside FileSystem._prepare() action graph.
An Install action is created for each Proxy action, as a result, when
a copy failed, only the related Proxy action is cancelled.

This patch adapts the code to have better error messages for that.

Change-Id: Icf5cf57e3385ec0854cb6d639db0dc319a71ce5e