Ticket #199: custom start/stop actions on target devices
This is a first implementation of device start/stop actions that are
defined in the model file as dev_action and associated with any targets
using dev_run=<action_alias>.

If dev_run is defined for a target, its device doesn't need to be
present on start, as the start command of the associated dev_action
will be executed prior to the target mount to allow the device to be
created. For stop, the Lustre target is first unmounted, and then the
device stop action is called.

Change-Id: I3910f4e4f077c7b25ba979ee78c450fddd221a33