Client.py: enhance multi-mount of same namespace
If several mount_path are defined and only one path is mounted, then
proc_lov_match can be true while the mount_path is not in /proc/mounts.

If proc_lov_match is true but the mount_path is not found in
/proc/mounts, then the state of this Client is OFFLINE.

Remove the up but not mounted = incoherent state test which cannot be
true anymore.

Change-Id: I3035d3f8d3d7649ac342c62595e4e4c572e50acc