StartTarget: add '-t lustre_tgt' to mount server targets
Shine now tries to mount both lustre_tgt and lustre filesystem types.
With this option, the mount command initially tries to execute
mount.lustre_tgt. When this command is not found, it fallbacks to
mount.lustre. With this option, Shine is both future proof while keeping
backward compatibility.

Note that lustre_tgt is set in the first place as it is more prone to
succeed in the future.

Co-authored-by: Rémi Palancher <remi@rackslab.io>
Change-Id: Ibed0c8936ca0c16da26cb48577973d1408db897d